+++ lib/cryptography/base_64.rb
@@ -15,8 +15,8 @@ class BASE64 < Cryptography
     prompt = TTY::Prompt.new
 
     choices = [
-      { name: 'Encode', value: -> { CryptographyAsciiArt.new('base64') && encode_base64 } },
-      { name: 'Decode', value: -> { CryptographyAsciiArt.new('base64') && decode_base64 } },
+      { name: 'Encrypt', value: -> { CryptographyAsciiArt.new('base64') && encrypt_base64 } },
+      { name: 'Decrypt', value: -> { CryptographyAsciiArt.new('base64') && decrypt_base64 } },
       { name: 'Go to previous menu', value: -> { Cryptography.new } },
       { name: 'Go to Main Menu', value: -> { Toolbox.new } },
       { name: 'Quit Program', value: -> { clear_terminal && exit } }
@@ -25,25 +25,25 @@ class BASE64 < Cryptography
     prompt.select('Please select your mode', choices, per_page: 5, cycle: true)
   end
 
-  def encode_base64
-    puts 'Please enter your data to be encoded'
+  def encrypt_base64
+    puts 'Please enter your data to be encrypted'
 
     data = gets.chomp
 
-    encoded = Base64.encode64(data)
+    encrypted = Base64.encode64(data)
 
-    print "\nYour result is: #{encoded.colorize(:green)}"
+    print "\nYour result is: #{encrypted.colorize(:green)}"
     quit_or_continue(BASE64)
   end
 
-  def decode_base64
-    puts 'Please enter your data to be encoded'
+  def decrypt_base64
+    puts 'Please enter your data to be encrypted'
 
     data = gets.chomp
 
-    decoded = Base64.decode64(data)
+    decrypted = Base64.decode64(data)
 
-    print "\nYour result is: #{decoded.colorize(:green)}"
+    print "\nYour result is: #{decrypted.colorize(:green)}"
     quit_or_continue(BASE64)
   end
 end